Only register signal handlers when Run accepts a context
Previously, if one of the WithSigCancel flavored Run methods were
called, the registered signal handler would prevent the normal SIGINT
interruption behavior even when the invoked command Run method didn't
accept a context argument to support cancellation. This change makes it
so that the signal handle is only registered if the underlying Run is
able to handle the cancellation.